Create create_delete_poly.c
This is the real meat of the updates. Dynamic Arrays!
The double poly[size] array found in the struct of the header has now been changed to double *poly. This lets us point poly to new memory locations which have been allocated to us by the malloc() function.

Read the comments for the full details.

TL;DRl
Every array created in the program, whether it be from the original two or a resultant of addition/division etc is now a dynamic array with properly allocated memory.